Refactor graphQL and concurrency functions api tests
